fix: field-data bugs (subcool/approach/delta-T sign, fault history func 0x41) and code review cleanups
